requirements.txt
djangorestframework-jwt
django-filter
djangorestframework
asgiref==3.4.1
captcha==0.3
Django==3.2.9
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
mysqlclient==2.0.3
operation==0.0.4
Pillow==8.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
